fix(crtp test): replace fragile analyzer-shell-out with a pure runtime regression
PR #81 review correctly noted the previous test was unreliable: the LHS annotation 'final MutationKey<_Bad, String, String> key = ...' itself trips MutationKey's own F-bound regardless of how MutationSerializableExtension is implemented, so the test would stay green even if the extension's safety regressed. The reviewer's suggested 'undefined_getter' assertion also doesn't work — the extension IS applicable to misdeclared subclasses (T resolves via the supertype to _Self) so no undefined-getter diagnostic ever fires.

The actual regression we're protecting against is the pre-#9 behaviour:  would throw TypeError at runtime when the F-bounded RequestType didn't match the runtime type. Replace the analyzer-shell-out test with a direct runtime test that calls .mutationKey on a misdeclared subclass and asserts it returns normally — this fails if the runtime cast (or similar unsafe coercion) is reintroduced.

- Drop test/fixtures/bad_mutation_serializable_crtp.dart (no longer needed).
- Rewrite test/src/models/serializable_crtp_misdeclaration_test.dart as a pure-Dart runtime test.
- Resolves both #85 (fixture rewrite) and #95 (analyzer-test tooling) — the simpler test obviates the tooling concerns.

Closes #85
Closes #95
