+++ test/src/models/serializable_crtp_misdeclaration_test.dart
@@ -1,30 +1,50 @@
-import 'dart:io';
-
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:typed_cached_query/src/errors/query_exception.dart';
+import 'package:typed_cached_query/src/models/serializable.dart';
+
+// Two declarations to set up a CRTP misdeclaration: _Self correctly self-types its F-bounded
+// RequestType, _Bad incorrectly types it as _Self.
+class _Self extends MutationSerializable<_Self, String, String> {
+  @override
+  String get keyGenerator => 'self';
+  @override
+  Future<dynamic> mutationFn() async => 'ok';
+  @override
+  String responseHandler(dynamic response) => response as String;
+  @override
+  OnErrorResults<_Self, String?> errorMapper(_Self request, String error, String? fallback) =>
+      OnErrorResults(request: request, error: MutationException(error, 500), fallback: fallback);
+}
+
+class _Bad extends MutationSerializable<_Self, String, String> {
+  @override
+  String get keyGenerator => 'bad';
+  @override
+  Future<dynamic> mutationFn() async => 'ok';
+  @override
+  String responseHandler(dynamic response) => response as String;
+  @override
+  OnErrorResults<_Self, String?> errorMapper(_Self request, String error, String? fallback) =>
+      OnErrorResults(request: request, error: MutationException(error, 500), fallback: fallback);
+}
+
 void main() {
-  test('CRTP misdeclaration is rejected by the analyzer (regression for #9)', () async {
-    // Run `dart analyze` against the deliberately-broken fixture and assert the expected
-    // diagnostics surface, proving the type-safe MutationSerializableExtension prevents what was
-    // previously a runtime cast failure.
-    final result = await Process.run(
-      'dart',
-      ['analyze', 'test/fixtures/bad_mutation_serializable_crtp.dart'],
-      workingDirectory: Directory.current.path,
-    );
+  test('misdeclared CRTP subclass does not crash when reading mutationKey (regression for #9)', () {
+    // The pre-#9 in-class getter performed `MutationKey(this as RequestType)`. With a misdeclared
+    // CRTP subclass (RequestType=_Self, instance type=_Bad), `this as _Self` throws a TypeError at
+    // runtime. The post-#9 extension resolves T via the declared supertype (T=_Self) and constructs
+    // a `MutationKey<_Self, String, String>` without any cast — no crash.
+    //
+    // If a regression reintroduces the runtime cast (or a similarly unsafe coercion), this test
+    // fails because reading `.mutationKey` will throw.
+    final bad = _Bad();
+    expect(() => bad.mutationKey, returnsNormally);
+  });
 
-    final output = '${result.stdout}\n${result.stderr}';
-
-    expect(result.exitCode, isNot(0), reason: 'dart analyze must fail on the misdeclared CRTP fixture\nOutput:\n$output');
-    expect(
-      output,
-      contains("doesn't conform to the bound"),
-      reason: 'Expected the F-bound violation diagnostic.\nOutput:\n$output',
-    );
-    expect(
-      output,
-      contains("can't be assigned to a variable of type"),
-      reason: 'Expected the assignment-type-mismatch diagnostic.\nOutput:\n$output',
-    );
-  }, timeout: const Timeout(Duration(seconds: 60)));
+  test('mutationKey on a correctly-declared CRTP subclass also does not crash', () {
+    // Sanity check — the well-formed case must continue to work.
+    final self = _Self();
+    expect(() => self.mutationKey, returnsNormally);
+  });
 }
